requirements.txt
Flask==3.0.2
numpy==1.26.4
Werkzeug==3.0.1
pycryptodome==3.20.0
scipy==1.12.0
python-magic==0.4.27 